+++ docs/reference/utilities.rst
@@ -4,5 +4,5 @@ Skafos Utilities
 Along with model version management, the Skafos SDK comes with additional tools to summarize your working environment
 and user privileges.
 
-.. automodule:: skafos.__init__
+.. automodule:: skafos.utilities
    :members:
